Add mukama as a scraping site
